Phase 4.6: Opportunity Map → Price Map (dense ranked list)
Tab + component renamed; the proportional price-ladder from 4.4 is
replaced with a dense ranked-list table.

Why: 4.4's ladder positioned cards at proportional y-from-price within
a fixed height. With clustered prices the cluster compressed then a
gap opened to the next cluster (image #4 — 200px of empty between
$33 and $30). Mathematically correct, visually awkward.

New layout:
- Single column, sorted price-desc, uniform 56px row height. No
  positioning math, no dead space.
- Each row: strength stripe · large bold price (left-anchored) ·
  thumb · brand + strength chip · reviews · rating · revenue/mo with
  a relative-size fill bar · external-link icon.
- Tier dividers (Premium / Mid / Value) appear at the 67th and 33rd
  price percentiles — a faint section header, not a giant gap.
  Skipped entirely when N < 4.
- Summary strip top-right: N competitors · Range $X–$Y · Median $Z.

Filters:
- Keep strength filter (All / Strong / Decent / Weak).
- Drop revenue tier filter — the relative revenue bar in each row
  now communicates revenue scale at a glance, the filter was redundant.

Plumbing:
- File renamed OpportunityMap.tsx → PriceMap.tsx; the rename shows
  as D+A in git because the rewrite is too thorough for git's
  similarity heuristic, but Phase 4.4's prior implementation lives in
  the git log on this branch under the old filename.
- Tab label, section header, and description copy updated.
- Internal activeTab key stays 'opportunity' to avoid migrating any
  persisted UI state. The user-facing surface says Price Map.
